fix: rename `prismicio.d.ts` before deleting `.slicemachine`

+++ packages/start-slicemachine/src/legacyMigrations/migrateAssets.ts
@@ -183,19 +183,22 @@ export const migrateAssets = async (
 			sentryErrorHandlers.node("migrateAssets", error);
 		}
 	} finally {
+		const prismicioTypesDTS = path.join(manager.cwd, "prismicio-types.d.ts")
+		if (!fsSync.existsSync(prismicioTypesDTS)) {
+			const sliceMachinePrismicioDTS = path.join(dotSlicemachine, "prismicio.d.ts");
+			if (fsSync.existsSync(sliceMachinePrismicioDTS)) {
+				fsSync.renameSync(sliceMachinePrismicioDTS, prismicioTypesDTS);
+			}
+			const prismicioDTS = path.join(manager.cwd, "prismicio.d.ts");
+			if (fsSync.existsSync(prismicioDTS)) {
+				fsSync.renameSync(prismicioDTS, prismicioTypesDTS);
+			}
+		}
+
 		const dotSlicemachine = createPathToDeprecatedLibrary(manager.cwd);
 		if (fsSync.existsSync(dotSlicemachine)) {
 			fsSync.rmSync(dotSlicemachine, { recursive: true });
 		}
 
-		const prismicioTypesDTS = path.join(manager.cwd, "prismicio-types.d.ts")
-		const sliceMachinePrismicioDTS = path.join(dotSlicemachine, "prismicio.d.ts");
-		if (fsSync.existsSync(sliceMachinePrismicioDTS)) {
-			fsSync.renameSync(sliceMachinePrismicioDTS, prismicioTypesDTS);
-		}
-		const prismicioDTS = path.join(manager.cwd, "prismicio.d.ts");
-		if (fsSync.existsSync(prismicioDTS)) {
-			fsSync.renameSync(prismicioDTS, prismicioTypesDTS);
-		}
 	}
 };
